requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
config==0.5.1
cors==1.0.1
Django==4.2.5
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
future==0.18.3
gevent==24.2.1
greenlet==3.0.3
idna==3.6
joblib==1.3.2
mysqlclient==2.2.4
nltk==3.8.1
numpy==1.26.4
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python-decouple==3.8
pytz==2024.1
regex==2023.12.25
requests==2.31.0
requests-file==2.0.0
setuptools==69.1.0
soupsieve==2.5
sqlparse==0.4.4
tldextract==5.1.1
tqdm==4.66.1
tzdata==2024.1
urllib3==2.2.0
zope.event==5.0
zope.interface==6.1
